SONAR-8857 remove redundant QProfileService class
