+++ server/sonar-server/src/main/java/org/sonar/server/qualityprofile/ws/ActivateRulesAction.java
@@ -24,10 +24,13 @@ import org.sonar.api.server.ServerSide;
 import org.sonar.api.server.ws.Request;
 import org.sonar.api.server.ws.Response;
 import org.sonar.api.server.ws.WebService;
+import org.sonar.server.organization.DefaultOrganizationProvider;
 import org.sonar.server.qualityprofile.BulkChangeResult;
-import org.sonar.server.qualityprofile.QProfileService;
+import org.sonar.server.qualityprofile.RuleActivator;
 import org.sonar.server.rule.ws.RuleQueryFactory;
+import org.sonar.server.user.UserSession;
 
+import static org.sonar.db.permission.OrganizationPermission.ADMINISTER_QUALITY_PROFILES;
 import static org.sonar.server.rule.ws.SearchAction.defineRuleSearchParameters;
 
 @ServerSide
@@ -38,12 +41,17 @@ public class ActivateRulesAction implements QProfileWsAction {
 
   public static final String BULK_ACTIVATE_ACTION = "activate_rules";
 
-  private final QProfileService profileService;
   private final RuleQueryFactory ruleQueryFactory;
+  private final UserSession userSession;
+  private final DefaultOrganizationProvider defaultOrganizationProvider;
+  private final RuleActivator ruleActivator;
 
-  public ActivateRulesAction(QProfileService profileService, RuleQueryFactory ruleQueryFactory) {
-    this.profileService = profileService;
+  public ActivateRulesAction(RuleQueryFactory ruleQueryFactory, UserSession userSession, DefaultOrganizationProvider defaultOrganizationProvider,
+    RuleActivator ruleActivator) {
     this.ruleQueryFactory = ruleQueryFactory;
+    this.userSession = userSession;
+    this.defaultOrganizationProvider = defaultOrganizationProvider;
+    this.ruleActivator = ruleActivator;
   }
 
   public void define(WebService.NewController controller) {
@@ -68,10 +76,15 @@ public class ActivateRulesAction implements QProfileWsAction {
 
   @Override
   public void handle(Request request, Response response) throws Exception {
-    BulkChangeResult result = profileService.bulkActivate(
-      ruleQueryFactory.createRuleQuery(request),
-      request.mandatoryParam(PROFILE_KEY),
-      request.param(SEVERITY));
+    verifyAdminPermission();
+    BulkChangeResult result = ruleActivator.bulkActivate(ruleQueryFactory.createRuleQuery(request), request.mandatoryParam(PROFILE_KEY), request.param(SEVERITY));
     BulkChangeWsResponse.writeResponse(result, response);
+  }
+
+  private void verifyAdminPermission() {
+    // FIXME check for the permission of the appropriate organization, not the default one
+    userSession
+      .checkLoggedIn()
+      .checkPermission(ADMINISTER_QUALITY_PROFILES, defaultOrganizationProvider.get().getUuid());
   }
 }
